src/Reader/requirements.txt
easyocr==1.7.1
pandas==2.2.3
Pillow==10.4.0
pydantic==2.9.2
python-dotenv==1.0.1
Requests==2.32.3
google-cloud-bigquery
google-api-core==2.11.1 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.17.3 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.3.3 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-storage==2.10.0 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.5.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.60.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud